+++ include/mqt-core/ir/QuantumComputation.hpp
@@ -324,8 +324,6 @@ public:
    */
   void measureAll(bool addBits = true, bool addBarrier = true);
 
-  void bridge(const Targets& targets);
-
   void reset(Qubit target);
   void reset(const Targets& targets);